% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{make_bubblemap}
\alias{make_bubblemap}
\alias{render_map}
\title{Create bubblemap}
\usage{
make_bubblemap(
  features,
  radius_col,
  radius_scale,
  fill_col,
  fill_scale,
  fill_opacity = 0.8,
  stroke = FALSE,
  ...
)

render_map(map, out_fn, bbox = NULL, timeout = 7200)
}
\arguments{
\item{features}{features in a spatial frame}

\item{radius_col}{string; name of the column to use for radius}

\item{radius_scale}{number; how much to scale the radius by}

\item{fill_col}{string; name of the column to use for colour}

\item{fill_scale}{Leaflet scale object (e.g. colorFactor/colorNumeric) or colour}

\item{fill_opacity}{number}

\item{stroke}{color}

\item{...}{additional parameters to pass to the addCircles() step}

\item{map}{Leaflet map object}

\item{out_fn}{string; filename to render to}

\item{bbox}{bounding box of area to render; if not provided, map will fit to
include all features}

\item{timeout}{integer; time to wait for rendering to finish. Defaults to
7200 (2 hours)}
}
\description{
Create a Leaflet bubblemap instance. The function will create a map
object. You can assign it to a variable and view the variable with the
build-in viewer, or you can use render_map() to generate a static image
of the map object.

Generates a static image from a map object using mapshot2. At high
resolutions and high object density, this can be a very expensive
operation.
}
\details{
You can also add legends or additional layers before rendering.

The rendering is done via a headless Chrome browser. See README for setup.
}
