% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{render_map}
\alias{render_map}
\title{Render map}
\usage{
render_map(
  map,
  out_fn,
  bbox = NULL,
  timeout = 7200,
  width = 3840,
  height = 2160
)
}
\arguments{
\item{map}{Leaflet map object}

\item{out_fn}{string; filename to render to}

\item{bbox}{bounding box of area to render; if not provided, map will fit to
include all features}

\item{timeout}{integer; time to wait for rendering to finish. Defaults to
7200 (2 hours)}
}
\description{
Generates a static image from a map object using mapshot2. At high
resolutions and high object density, this can be a very expensive
operation.
}
\details{
The rendering is done via a headless Chrome browser. See README for setup.
}
