% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{squeeze_bbox}
\alias{squeeze_bbox}
\title{Squeeze bounding box}
\usage{
squeeze_bbox(
  features,
  capture_targ = 0.9995,
  step_size = 0.025,
  max_steps = 50,
  sample_rate = 0.5
)
}
\arguments{
\item{features}{features in a spatial frame}

\item{capture_targ}{aim for a box that fits this proportion of features}

\item{step_size}{shrink the box by this much every step}

\item{max_steps}{maximum numbers of steps to attempt before giving up}

\item{sample_rate}{apply on a random sample of the features instead}
}
\description{
Generate a more efficient bounding box by dropping outliers
}
